+++ src/Controller/Admin/BoardMemberCrudController.php
@@ -21,7 +21,7 @@ class BoardMemberCrudController extends AbstractCrudController
             ->setPageTitle('new', 'Add boardmember')
             ->setEntityLabelInSingular('Boardmember')
             ->setEntityLabelInPlural('Boardmembers')
-            ->setSearchFields(['name', 'board.name'])
+            ->setSearchFields(['name'])
             ->setDefaultSort(['name' => 'ASC'])
         ;
     }
@@ -29,9 +29,11 @@ class BoardMemberCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         yield TextField::new('name', 'Name');
-        yield AssociationField::new('board', 'Board')
-            ->setRequired(true);
         yield AssociationField::new('municipality', 'Municipality')
             ->setRequired(true);
+        yield AssociationField::new('boardRoles', 'BoardRole')
+            ->setFormTypeOptions([
+                'by_reference' => false,
+            ]);
     }
 }