+++ CHANGELOG.md
@@ -41,6 +41,7 @@ MunicipalityListener, BoardListener and their tests.
 - [TVIST1-186](https://jira.itkdev.dk/browse/TVIST1-186): Added ability to copy documents to other cases.
 - [TVIST1-126](https://jira.itkdev.dk/browse/TVIST1-126): Adding workflows for cases.
 - [TVIST1-227](https://jira.itkdev.dk/browse/TVIST1-227): Added ability to create, edit and remove notes.
+- [TVIST1-322](https://jira.itkdev.dk/browse/TVIST1-322): Replaced SubBoard entity by BoardRole entity.
 
 ### Changed
 - [NSEK-138](https://jira.itkdev.dk/browse/NSEK-138): Updating Cypress to 6.5.0 so the same version is used project wide.